backend/requirements.txt
amqp==5.0.2
asgiref==3.2.7
billiard==3.6.3.0
boto3==1.16.28
botocore==1.19.28
celery==5.0.2
cffi==1.14.0
click-didyoumean==0.0.3
click-repl==0.1.6
click==7.1.2
coverage==5.3
cryptography==3.2
cycler==0.10.0
django-celery-beat==2.1.0
django-cors-headers==3.5.0
django-filter==2.2.0
django-storages==1.10.1
django-timezone-field==4.1.1
django==3.0.7
djangorestframework-simplejwt==4.6.0
djangorestframework==3.11.0
importlib-metadata==1.6.0
jmespath==0.10.0
kiwisolver==1.3.1
kombu==5.0.2
markdown==3.2.2
matplotlib==3.3.3
model-bakery==1.2.1
numpy==1.19.4
pillow==7.1.2
prompt-toolkit==3.0.8
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pyjwt==1.7.1
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
redis==3.5.3
reportlab==3.5.56
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.26.2; python_version != '3.4'
vine==5.0.0
wcwidth==0.2.5
zipp==3.1.0
